chore(copilot): trim unused server mode
